requirements.txt
# Core dependencies
neo4j==5.15.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0

# Data processing
numpy==1.26.0
scikit-learn==1.3.2
datasets==2.16.1

# Vector databases
chromadb==1.5.5

# LLM integration
openai==1.3.9
ollama==0.6.1

# Utilities
tqdm==4.66.1
pyyaml==6.0.1

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
